requirements.txt

fastapi
uvicorn[standard]
python-multipart
aiofiles
requests
pydantic
starlette
h11
scipy
wandb

torch==1.11.0+cpu
torchvision==0.12.0+cpu
-f https://download.pytorch.org/whl/cpu/torch_stable.html

albumentations==1.0.3
imgaug==0.4.0
opencv-python==4.5.1.48
opencv-python-headless==4.5.1.48
Pillow==8.2.0
imageio==2.9.0
tifffile==2022.8.12
pywavelets==1.3.0
shapely==1.8.4

numpy==1.19.2
pandas==1.2.3
scikit-learn==1.0.2
scikit-image==0.18.1
matplotlib==3.4.1
seaborn==0.11.2
tqdm==4.62.3
joblib==1.1.0

boto3==1.24.56
botocore==1.27.56
s3transfer==0.6.0

pyyaml==5.4.1
typing_extensions==4.3.0
python-dateutil==2.8.2
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
pyparsing==3.0.9
packaging==21.3
cycler==0.11.0
kiwisolver==1.4.4
retrying==1.3.3
urllib3==1.25.11
certifi==2022.6.15
charset-normalizer==2.0.4
idna==3.3
zipp==3.8.1

cffi==1.15.1
pycparser==2.21
cryptography==37.0.1
pyopenssl==22.0.0
pysocks==1.7.1
win_inet_pton==1.1.0

hydra-core==1.1.0
omegaconf==2.1.2
markupsafe==2.1.1
Jinja2==3.1.2
click==8.1.3
networkx==2.8.6
threadpoolctl==3.1.0

git+https://github.com/Alexandre-Delplanque/HerdNet.git
